Corrected model default values
When providing a callable as an default value, if a call to the callable is provided as default
value, the value of the callable when called at the time of the server restart is
used. This is a problem in timezone.now(), since the datetime of the server restart
will be used for all default values. Thus, remove the calls to the callable, and just
provide the callable.